feat: finish jobs

+++ dashboard/src/features/job-oppenings/pages/qualified-oppening-page.tsx
@@ -5,21 +5,34 @@ import {
   TableHead,
   TableHeader,
   TableRow,
-} from "@/shared/components/ui/table"
-import { useFetchJobOppenings } from "../hooks/use-fetch-many-job-oppenings"
-import { Button } from "@/shared/components/ui/button"
-import { useDeleteJobOppening } from "../hooks/use-delete-job-oppening"
-import { Trash } from "lucide-react"
-import UpdateQualifiendOppening from "../components/update-qualifiend-oppening"
+} from "@/shared/components/ui/table";
+import { useFetchJobOppenings } from "../hooks/use-fetch-many-job-oppenings";
+import { Button } from "@/shared/components/ui/button";
+import { useDeleteJobOppening } from "../hooks/use-delete-job-oppening";
+import { Trash } from "lucide-react";
+import UpdateQualifiendOppening from "../components/update-qualifiend-oppening";
+import EmptyState from "@/shared/components/common/empty-state";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/shared/components/ui/alert-dialog";
+import { convertToDate } from "@/shared/utils/convert-date";
 
 export default function QualifiedOppeningPage() {
-  const data = useFetchJobOppenings()
-  const deleteHook = useDeleteJobOppening()
+  const data = useFetchJobOppenings();
+  const deleteHook = useDeleteJobOppening();
 
-  const handleDelete = async (id: string) => await deleteHook.mutateAsync(id)
+  const handleDelete = async (id: string) => await deleteHook.mutateAsync(id);
 
   if (data.isLoading) {
-    return <div>Carregando...</div>
+    return <div>Carregando...</div>;
   }
 
   if (data.isError) {
@@ -28,24 +41,25 @@ export default function QualifiedOppeningPage() {
         Erro ao carregar dados.
         <Button onClick={() => data.refetch()}>Tentar novamente</Button>
       </div>
-    )
+    );
   }
 
   if (!data.data?.data.length) {
-    return <div>Não há vagas disponíveis.</div>
+    return <EmptyState message="Nenhuma vaga disponível" />;
   }
 
-  const jobs = data.data.data
+  const jobs = data.data.data;
 
   return (
     <div>
       <Table className="bg-white">
         <TableHeader className="bg-LT-GRAY-200/20">
           <TableRow>
-            <TableHead>Função</TableHead>
-            <TableHead>Departamento</TableHead>
-            <TableHead>Quantidade</TableHead>
-            <TableHead>Ações</TableHead>
+            <TableHead className="md:w-[400px]">Função</TableHead>
+            <TableHead className="md:w-[400px]">Departamento</TableHead>
+            <TableHead className="md:w-[400px]">Quantidade</TableHead>
+            <TableHead className="md:w-[400px]">Data de criação</TableHead>
+            <TableHead className="md:w-[400px]">Ações</TableHead>
           </TableRow>
         </TableHeader>
         <TableBody>
@@ -54,21 +68,40 @@ export default function QualifiedOppeningPage() {
               <TableCell>{job.title}</TableCell>
               <TableCell>{job.department}</TableCell>
               <TableCell>{job.quantity}</TableCell>
+              <TableCell>{convertToDate(job.createdAt.toString())}</TableCell>
               <TableCell className="flex items-center gap-2">
                 <UpdateQualifiendOppening job={job} />
-                <Button
-                  size={"icon"}
-                  disabled={deleteHook.isPending}
-                  onClick={() => handleDelete(job.id)}
-                  variant={"destructive"}
-                >
-                  <Trash />
-                </Button>
+                <AlertDialog>
+                  <AlertDialogTrigger asChild>
+                    <Button size={"icon"} variant={"destructive"}>
+                      <Trash />
+                    </Button>
+                  </AlertDialogTrigger>
+                  <AlertDialogContent>
+                    <AlertDialogHeader>
+                      <AlertDialogTitle>
+                        Tens a certeza que pretendes eliminar esta vaga?
+                      </AlertDialogTitle>
+                      <AlertDialogDescription>
+                        Esta ação não é reversível
+                      </AlertDialogDescription>
+                    </AlertDialogHeader>
+                    <AlertDialogFooter>
+                      <AlertDialogCancel>Cancelar</AlertDialogCancel>
+                      <AlertDialogAction
+                        disabled={deleteHook.isPending}
+                        onClick={() => handleDelete(job.id)}
+                      >
+                        Remover
+                      </AlertDialogAction>
+                    </AlertDialogFooter>
+                  </AlertDialogContent>
+                </AlertDialog>
               </TableCell>
             </TableRow>
           ))}
         </TableBody>
       </Table>
     </div>
-  )
+  );
 }
